Refactoring "app" name
